requirements.txt
absl-py==0.13.0
colorlog==6.5.0
huggingface-hub==0.4.0
hydra-colorlog==1.1.0
hydra-core==1.1.1
hydra-submitit-launcher==1.1.6
jsonlines==2.0.0
munkres==1.1.4
natsort==8.1.0
omegaconf==2.1.1
pyaml-env==1.1.1
pygame==2.0.1
pytorch-lightning==1.5.0
PyYAML==5.4.1
regex==2022.3.15
rich==10.12.0
scipy==1.6.0
seaborn==0.11.2
submitit==1.4.0
torch==1.10.0
torchmetrics==0.5.1
torchvision==0.11.1
torchviz==0.0.2
transformers==4.17.0
virtualenv==20.0.6
Wand==0.6.7
wandb==0.12.10